Implement sequences
Define a sequence of conditions to be satisfied
in chronological order, from the first to the last.

Query is like:

{"sequence": [{"sub-fen": "8/3q4/8/8/8/8/8/8"},
              {"sub-fen": "2q5/8/8/8/8/8/8/R6R"}]}

And result is like:

{
    "moves": 3406395,
    "match count": 10,
    "moves/second": 7536272,
    "processing time (ms)": 452,
    "matches":
    [
        { "ofs": 940231, "ply": [22, 28, ] },
        { "ofs": 2274077, "ply": [42, 46, ] },
        { "ofs": 7576222, "ply": [24, 30, ] },
        { "ofs": 8149624, "ply": [34, 48, ] },
        { "ofs": 13483146, "ply": [20, 26, ] },
        { "ofs": 15434253, "ply": [46, 48, ] },
        { "ofs": 15882036, "ply": [10, 18, ] },
        { "ofs": 19714828, "ply": [40, 46, ] },
        { "ofs": 23825029, "ply": [46, 56, ] },
        { "ofs": 26171395, "ply": [28, 48, ] },
    ]
}

Note that the "ply" tag becomes a list of matching plies,
one for each condition. For instance in case of the game
at offset 13483146, the first sub-fen was found at ply 20
and the second at ply 26.