First cut at implementing streak sequence
As a side effect, to simplify the code, I
have changed the semantic of normal sequence
so that once a condition is satisfied, the next
one will be tested starting from the next move
and no more from the same move again.

This should be a non-issue in real cases.