Use SHA1 to verify test queries
Instead of checking just the number of matches for each query, now
check against a SHA1 hash of the full match info, including offsets
and plies.

This is more powerful because we can spot a wrong result
even if the number of matches remains the same but it changes
just one offset or one ply.

Also run pep8 on it