Revert new move_rule logic
Introduces inconsistencies in the search. See https://github.com/mcostalba/scoutfish/issues/41

Fixing them requires even more tweaks and logic. Just revert to the simple convention used
before.

Now 'ply' is no more what players may expect, but at least it is consistent across all
the rules and conditions. It is up to the GUI to 'fix' it, indeed it is not even a
value visible by the user, it is assumed to be consumed by the GUI.