Add support for passing the starting offset
When making an index it is now possible to pass the starting offset, that
will be used instead of 0.

This is useful when appending a new pgn file to an exsisting one. Suppose we have ann already exsisting
db.pgn and corresponding db.scout and we want to add new.pgn and update db.scout withouth rebuilding
the index. In this case, first create new.scout as follow:

   scoutfish make new.pgn startOffset

Where startOffset is set to size of the old db.pgn. Then just append new.pgn to db.pgn and append
new.scout to db.scout. That's all!

Some renaming while there.

This should fix https://github.com/mcostalba/scoutfish/issues/28